pyproject.toml
[project]
name = "lean-alpha"
version = "0.2.0"
description = "Lean Agentic Alpha — A1 coordinator with specialist subagents (Claude Agent SDK)"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "anthropic>=0.92.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
    "markdown>=3.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
]

[project.scripts]
lean-alpha = "lean_alpha.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["lean_alpha*"]
